requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-modeltranslation==0.17.3
Pillow==8.4.0
six==1.16.0
sqlparse==0.4.2
